feat: support `ask` option to close_buffer + smart_close fix (#40)
* actually support `ask` option to close_buffer

* fix: avoid closing tab when not necessary

When we close the last buf for a file but there are still other bufs
opened in the same tab we should not kill the entire tab.
Instead first switch to the previous buffer in the tab and then delete
the buf being closed.
This is the same way it is handled when the buffer is opened in other
tabs and there are other bufs in the current tab.